+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Fixed
+- [#38](https://github.com/BixData/stuart-ml/issues/38) Vectors.dense(...) with varargs error: no table.pack() function (eLua interop)
+
 ## [1.0.0] - 2018-11-08
 ### Changed
 - Upgrade to Stuart 1.0.0 with changes to class framework